refactor(test): PathFinderTest에서 필요변수 선언부분 분리
