+++ src/test/java/nextstep/subway/path/domain/PathFinderTest.java
@@ -6,47 +6,11 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.Arrays;
 import java.util.List;
 import nextstep.subway.line.domain.Line;
-import nextstep.subway.line.domain.Section;
 import nextstep.subway.station.domain.Station;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class PathFinderTest {
-
-    private Station 강남역;
-    private Station 양재역;
-    private Station 교대역;
-    private Station 남부터미널역;
-    private Line 신분당선;
-    private Line _2호선;
-    private Line _3호선;
-    private List<Line> 노선목록;
-
-    private Section 구간;
-
-    /**
-     * 교대역    --- *2호선* ---   강남역
-     * |                        |
-     * *3호선*                   *신분당선*
-     * |                        |
-     * 남부터미널역  --- *3호선* ---   양재
-     */
-    @BeforeEach
-    public void setUp() {
-        강남역 = new Station("강남역");
-        양재역 = new Station("양재역");
-        교대역 = new Station("교대역");
-        남부터미널역 = new Station("남부터미널역");
-
-        신분당선 = new Line("신분당선", "red", 강남역, 양재역, 10);
-        _2호선 = new Line("2호선", "green", 교대역, 강남역, 10);
-        _3호선 = new Line("3호선", "orange", 교대역, 양재역, 5);
-        노선목록 = Arrays.asList(신분당선, _2호선, _3호선);
-
-        구간 = new Section(_3호선, 교대역, 남부터미널역, 3);
-        _3호선.addSection(구간);
-    }
+class PathFinderTest extends PathTestFixture {
 
     @DisplayName("환승하지 않는 지하철 경로 조회")
     @Test
